[expo-cli] move android submit command logic to a separate file
