+++ packages/expo-cli/src/commands/upload/submission-service/types.ts
@@ -1,0 +1,13 @@
+export interface SubmissionContext<T extends SubmitCommandOptions> {
+  projectDir: string;
+  options: T;
+}
+
+export interface SubmitCommandOptions {
+  latest?: boolean;
+  id?: string;
+  path?: string;
+  url?: string;
+  verbose?: boolean;
+  useSubmissionService?: boolean;
+}
